refactoring: Introduce InspectrContext
