+++ src/components/InspectrApp.jsx
@@ -5,32 +5,35 @@ import RequestList from './RequestList';
 import RequestDetailsPanel from './RequestDetailsPanel';
 import SettingsPanel from './SettingsPanel';
 import eventDB from '../utils/eventDB';
-import ToastNotification from './ToastNotification.jsx';
 import useInspectrRouter from '../hooks/useInspectrRouter.jsx';
+import { useInspectr } from '../context/InspectrContext';
+import { deleteAllOperationsApi as deleteAllOperationsApiUtil, deleteOperationApi as deleteOperationApiUtil } from '../utils/inspectrSdk';
 
-const isLocalhost = typeof window !== 'undefined' && window.location.hostname === 'localhost';
 const debugMode = typeof window !== 'undefined' && localStorage.getItem('debug') === 'true';
 
 if (debugMode) {
   console.log('[Inspectr] Debug Mode enabled');
 }
 
-const InspectrApp = ({ apiEndpoint: initialApiEndpoint = '/api' }) => {
-  const [apiEndpoint, setApiEndpoint] = useState(initialApiEndpoint);
-
-  const [connectionStatus, setConnectionStatus] = useState('disconnected');
-
-  // Registration details state.
-  const [sseEndpoint, setSseEndpoint] = useState('');
-  const [channelCode, setChannelCode] = useState('');
-  const [channel, setChannel] = useState('');
-  const [token, setToken] = useState('');
-  const [expires, setExpires] = useState('');
-
-  // Track initialization state
-  const [isInitialized, setIsInitialized] = useState(false);
-
-  const [toast, setToast] = useState(null);
+const InspectrApp = () => {
+  // Get all the shared state from context
+  const { 
+    apiEndpoint,
+    setApiEndpoint,
+    connectionStatus,
+    setConnectionStatus,
+    sseEndpoint,
+    channelCode,
+    setChannelCode,
+    channel,
+    setChannel,
+    token,
+    reRegistrationFailedRef,
+    handleRegister,
+    attemptReRegistration,
+    toast,
+    setToast
+  } = useInspectr();
 
   const pageSize = 100;
   const [page, setPage] = useState(1);
@@ -41,10 +44,6 @@ const InspectrApp = ({ apiEndpoint: initialApiEndpoint = '/api' }) => {
 
   // SSE connection reference
   const wasConnectedRef = useRef(false);
-  const registrationRetryCountRef = useRef(0);
-  const reRegistrationFailedRef = useRef(false);
-  const maxRegistrationRetries = 6;
-  const retryDelay = 5000; // milliseconds
 
   // Live query: get the events for the current page.
   const { results: operations = [], totalCount = 0 } =
@@ -66,219 +65,6 @@ const InspectrApp = ({ apiEndpoint: initialApiEndpoint = '/api' }) => {
   const { selectedOperation, currentTab, handleSelect, handleTabChange, clearSelection } =
     useInspectrRouter(operations);
 
-  // Load credentials from URL query parameters
-  const loadCredentialsFromQueryParams = () => {
-    const urlParams = new URLSearchParams(window.location.search);
-    const queryChannelCode = urlParams.get('channelCode');
-    const queryChannel = urlParams.get('channel');
-    const queryToken = urlParams.get('token');
-
-    if (queryChannelCode || queryChannel || queryToken) {
-      console.log('🔍 Found credentials in query params');
-      if (queryChannelCode) {
-        setChannelCode(queryChannelCode);
-        localStorage.setItem('channelCode', queryChannelCode);
-      }
-      if (queryChannel) {
-        setChannel(queryChannel);
-        localStorage.setItem('channel', queryChannel);
-      }
-      if (queryToken) {
-        setToken(queryToken);
-        localStorage.setItem('token', queryToken);
-      }
-      // Update the URL without reloading the page.
-      window.history.replaceState(
-        {},
-        '',
-        `${window.location.pathname}${urlParams.toString() ? '?' + urlParams.toString() : ''}`
-      );
-      return true;
-    }
-    return false;
-  };
-
-  // Load credentials from local storage
-  const loadCredentialsFromLocalStorage = () => {
-    const storedChannelCode = localStorage.getItem('channelCode');
-    const storedChannel = localStorage.getItem('channel');
-    const storedToken = localStorage.getItem('token');
-
-    if (storedChannelCode && storedChannel && storedToken) {
-      console.log('✅ Using stored credentials from localStorage');
-      setChannelCode(storedChannelCode);
-      setChannel(storedChannel);
-      setToken(storedToken);
-      return true;
-    }
-    return false;
-  };
-
-  // Load credentials from REST API
-  const loadCredentialsFromApi = async () => {
-    console.log('🔄 Fetching /app/config (Localhost, No credentials found)');
-    try {
-      const result = await getConfigApi();
-      if (result?.token && result?.sse_endpoint && result?.channel_code) {
-        console.log('✅ Loaded from /app/config:', result);
-        setChannelCode(result.channel_code);
-        localStorage.setItem('channelCode', result.channel_code);
-        setChannel(result.channel);
-        localStorage.setItem('channel', result.channel);
-        setToken(result.token);
-        localStorage.setItem('token', result.token);
-        setSseEndpoint(result.sse_endpoint);
-        localStorage.setItem('sseEndpoint', result.sse_endpoint);
-        localStorage.setItem('ingressEndpoint', result.ingress_endpoint);
-        localStorage.setItem('proxyEndpoint', result.proxy_endpoint);
-        localStorage.setItem('expose', result.expose ? 'true' : 'false');
-      }
-    } catch (err) {
-      console.error('❌ Failed to load /app/config:', err);
-    }
-  };
-
-  /**
-   * 🏁 Step 1: Load credentials from localStorage on mount
-   */
-  useEffect(() => {
-    const loadCredentials = async () => {
-      if (typeof window === 'undefined') return;
-
-      if (!loadCredentialsFromQueryParams()) {
-        if (!loadCredentialsFromLocalStorage() && isLocalhost) {
-          await loadCredentialsFromApi();
-        }
-      }
-
-      setIsInitialized(true);
-    };
-
-    loadCredentials();
-  }, []);
-
-  /**
-   * 🏁 Step 2: When `isInitialized` is true, register using stored credentials
-   */
-  useEffect(() => {
-    if (!isInitialized) return;
-    console.log('🔄 Auto-registering with stored credentials:', channel, channelCode);
-
-    if (channel && channelCode) {
-      handleRegister(channelCode, channel);
-    } else {
-      console.log('⚠️ Missing credentials for auto-registration, skipping.');
-    }
-  }, [isInitialized, channel, channelCode]);
-
-  // Registration handler.
-  const handleRegister = async (
-    newChannelCode = channelCode,
-    newChannel = channel,
-    newToken = token,
-    showNotification
-  ) => {
-    try {
-      // Construct request body
-      let requestBody = {};
-      if (newChannelCode && newChannel) {
-        requestBody = { channel: newChannel, channel_code: newChannelCode };
-      } else if (newToken) {
-        requestBody = { token: newToken };
-      }
-
-      console.log('📤 Registering with:', requestBody);
-
-      // Register App with Inspectr
-      const result = await registerApi(requestBody);
-
-      if (result?.token && result?.sse_endpoint && result?.channel_code) {
-        console.log('✅ Registration successful');
-        registrationRetryCountRef.current = 0;
-        reRegistrationFailedRef.current = false;
-
-        setChannelCode(result.channel_code);
-        localStorage.setItem('channelCode', result.channel_code);
-        setChannel(result.channel);
-        localStorage.setItem('channel', result.channel);
-        setToken(result.token);
-        localStorage.setItem('token', result.token);
-        setExpires(result.expires);
-        localStorage.setItem('expires', result.expires);
-        setSseEndpoint(result.sse_endpoint);
-        localStorage.setItem('sseEndpoint', result.sse_endpoint);
-        localStorage.setItem('ingressEndpoint', result.ingress_endpoint);
-        localStorage.setItem('proxyEndpoint', result.proxy_endpoint);
-        localStorage.setItem('expose', result.expose ? 'true' : 'false');
-
-        if (showNotification) {
-          setToast({
-            message: 'Registration Successful',
-            subMessage: 'Your channel and access code have been registered.'
-          });
-        }
-        // Update the connection status.
-        setConnectionStatus('connected');
-        return true;
-      } else {
-        console.error('❌ Registration failed:', result);
-        setToast({
-          message: 'Registration Failed',
-          subMessage: 'Please check your channel and access code.',
-          type: 'error'
-        });
-        throw new Error('Registration failed');
-      }
-    } catch (error) {
-      console.error('❌ Registration error:', error);
-      setToast({
-        message: 'Registration Error',
-        subMessage: 'An error occurred during registration.',
-        type: 'error'
-      });
-      throw error;
-    }
-  };
-
-  const attemptReRegistration = () => {
-    setConnectionStatus('reconnecting');
-    if (registrationRetryCountRef.current < maxRegistrationRetries) {
-      registrationRetryCountRef.current += 1;
-      console.log(
-        `🔄 Attempting re-registration (${registrationRetryCountRef.current}/${maxRegistrationRetries}) in ${
-          retryDelay / 1000
-        } seconds...`
-      );
-      setTimeout(async () => {
-        try {
-          await handleRegister();
-        } catch (error) {
-          console.log('❌ Re-registration attempt failed:', error);
-          // Try again recursively.
-          attemptReRegistration();
-        }
-      }, retryDelay);
-    } else {
-      console.log(`❌ Re-registration failed after ${maxRegistrationRetries} attempts. Giving up.`);
-      registrationRetryCountRef.current = 0;
-      reRegistrationFailedRef.current = true; // Prevent further attempts
-      // Set final connection status to "disconnected"
-      setConnectionStatus('disconnected');
-    }
-  };
-
-  // Automatically trigger registration when a channel is present and token is not yet set.
-  // useEffect(() => {
-  //   if (localStorageLoaded) {
-  //     if (channel && channelCode) {
-  //       console.log('Automatically trigger reregistration', channel, channelCode);
-  //       handleRegister(channelCode, channel);
-  //     } else {
-  //       console.log('Automatically trigger new registration');
-  //       handleRegister();
-  //     }
-  //   }
-  // }, [localStorageLoaded, channel, channelCode]);
 
   // Connect to SSE when the component mounts.
   useEffect(() => {
@@ -333,7 +119,7 @@ const InspectrApp = ({ apiEndpoint: initialApiEndpoint = '/api' }) => {
       }
 
       // Start the re-registration retry loop if not already in progress.
-      if (!reRegistrationFailedRef.current && registrationRetryCountRef.current === 0) {
+      if (!reRegistrationFailedRef.current) {
         console.log('🔄 Starting re-registration retry loop due to SSE error.');
         attemptReRegistration();
       }
@@ -345,7 +131,7 @@ const InspectrApp = ({ apiEndpoint: initialApiEndpoint = '/api' }) => {
       eventSource.close();
       setConnectionStatus('disconnected');
     };
-  }, [sseEndpoint]); // Run only once on mount
+  }, [sseEndpoint, token, setConnectionStatus, reRegistrationFailedRef, attemptReRegistration]); // Run when these dependencies change
 
   // If no operation is selected but there are operations, select the first one.
   useEffect(() => {
@@ -441,16 +227,7 @@ const InspectrApp = ({ apiEndpoint: initialApiEndpoint = '/api' }) => {
   // Delete all operations via REST API
   const deleteAllOperationsApi = async () => {
     try {
-      const response = await fetch(`${apiEndpoint}/operations`, {
-        method: 'DELETE',
-        headers: {
-          'Content-Type': 'application/json',
-          'inspectr-client': 'inspectr-app'
-        }
-      });
-      if (!response.ok) {
-        throw new Error('Failed to delete all operations');
-      }
+      await deleteAllOperationsApiUtil(apiEndpoint);
     } catch (error) {
       console.error('Error deleting all operations:', error);
       setToast({
@@ -464,16 +241,7 @@ const InspectrApp = ({ apiEndpoint: initialApiEndpoint = '/api' }) => {
   // Delete a single operation via REST API
   const deleteOperationApi = async (id) => {
     try {
-      const response = await fetch(`${apiEndpoint}/operations/${id}`, {
-        method: 'DELETE',
-        headers: {
-          'Content-Type': 'application/json',
-          'inspectr-client': 'inspectr-app'
-        }
-      });
-      if (!response.ok) {
-        throw new Error(`Failed to delete operation ${id}`);
-      }
+      await deleteOperationApiUtil(apiEndpoint, id);
     } catch (error) {
       console.error(`Error deleting operation ${id}:`, error);
       setToast({
@@ -484,30 +252,6 @@ const InspectrApp = ({ apiEndpoint: initialApiEndpoint = '/api' }) => {
     }
   };
 
-  // Register Inspectr App via REST API
-  const registerApi = async (requestBody) => {
-    const response = await fetch(`${apiEndpoint}/register`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        'inspectr-client': 'inspectr-app'
-      },
-      body: JSON.stringify(requestBody)
-    });
-    if (!response.ok) {
-      throw new Error(`Registration API call failed with status ${response.status}`);
-    }
-    return await response.json();
-  };
-
-  // Get local Inspectr App config via REST API
-  const getConfigApi = async () => {
-    const response = await fetch('/app/config');
-    if (!response.ok) {
-      throw new Error(`Failed to fetch config: ${response.status}`);
-    }
-    return await response.json();
-  };
 
   return (
     <div
@@ -549,25 +293,7 @@ const InspectrApp = ({ apiEndpoint: initialApiEndpoint = '/api' }) => {
         </div>
       </div>
       {/* Bottom Panel */}
-      <SettingsPanel
-        apiEndpoint={apiEndpoint}
-        setApiEndpoint={setApiEndpoint}
-        connectionStatus={connectionStatus}
-        channelCode={channelCode}
-        setChannelCode={setChannelCode}
-        channel={channel}
-        setChannel={setChannel}
-        onRegister={handleRegister}
-        onReconnect={attemptReRegistration}
-      />
-      {toast && (
-        <ToastNotification
-          message={toast.message}
-          subMessage={toast.subMessage}
-          type={toast.type}
-          onClose={() => setToast(null)}
-        />
-      )}
+      <SettingsPanel />
     </div>
   );
 };
