+++ src/components/Workspace.jsx
@@ -5,6 +5,8 @@ import DashBoardApp from './DashBoardApp.jsx';
 import InspectrApp from './InspectrApp.jsx';
 import SettingsApp from './SettingsApp.jsx';
 import useHashRouter from '../hooks/useHashRouter.jsx';
+import ToastNotification from './ToastNotification.jsx';
+import { InspectrProvider, useInspectr } from '../context/InspectrContext';
 
 function classNames(...classes) {
   return classes.filter(Boolean).join(' ');
@@ -38,50 +40,71 @@ export default function Workspace() {
   const ActiveComponent = currentNav.component;
 
   return (
-    <div className="">
-      <div className="border-b border-tremor-border dark:border-dark-tremor-border relative h-full overflow-hidden bg-gray-50 dark:bg-dark-tremor-background-subtle">
-        <div className="px-4 sm:px-6 lg:px-8">
-          <div className="overflow flex h-16 sm:space-x-7">
-            <div className="hidden shrink-0 sm:flex sm:items-center">
-              <a href="/" className="p-1.5">
-                <Logo
-                  className="size-4 shrink-0 w-8 h-8 text-tremor-content-strong dark:text-dark-tremor-content-strong"
-                  aria-hidden={true}
-                />
-              </a>
+    <InspectrProvider>
+      <div className="">
+        <div className="border-b border-tremor-border dark:border-dark-tremor-border relative h-full overflow-hidden bg-gray-50 dark:bg-dark-tremor-background-subtle">
+          <div className="px-4 sm:px-6 lg:px-8">
+            <div className="overflow flex h-16 sm:space-x-7">
+              <div className="hidden shrink-0 sm:flex sm:items-center">
+                <a href="/" className="p-1.5">
+                  <Logo
+                    className="size-4 shrink-0 w-8 h-8 text-tremor-content-strong dark:text-dark-tremor-content-strong"
+                    aria-hidden={true}
+                  />
+                </a>
+              </div>
+              <nav className="-mb-px flex space-x-6" aria-label="Tabs">
+                {navigation.map((navItem) => (
+                  <button
+                    key={navItem.slug}
+                    onClick={() => handleTabClick(navItem)}
+                    className={classNames(
+                      navItem.slug === currentNav.slug
+                        ? 'dark:text-tremor-dark-brand border-tremor-brand text-tremor-brand'
+                        : 'border-transparent text-tremor-content-emphasis hover:border-tremor-content-subtle hover:text-tremor-content-strong dark:text-dark-tremor-content-emphasis hover:dark:border-dark-tremor-content-subtle hover:dark:text-dark-tremor-content-strong',
+                      'inline-flex items-center whitespace-nowrap border-b-2 px-2 text-tremor-default font-medium'
+                    )}
+                    aria-current={navItem.name === currentTab.name ? 'page' : undefined}
+                  >
+                    {navItem.name}
+                  </button>
+                ))}
+              </nav>
             </div>
-            <nav className="-mb-px flex space-x-6" aria-label="Tabs">
-              {navigation.map((navItem) => (
-                <button
-                  key={navItem.slug}
-                  onClick={() => handleTabClick(navItem)}
-                  className={classNames(
-                    navItem.slug === currentNav.slug
-                      ? 'dark:text-tremor-dark-brand border-tremor-brand text-tremor-brand'
-                      : 'border-transparent text-tremor-content-emphasis hover:border-tremor-content-subtle hover:text-tremor-content-strong dark:text-dark-tremor-content-emphasis hover:dark:border-dark-tremor-content-subtle hover:dark:text-dark-tremor-content-strong',
-                    'inline-flex items-center whitespace-nowrap border-b-2 px-2 text-tremor-default font-medium'
-                  )}
-                  aria-current={navItem.name === currentTab.name ? 'page' : undefined}
-                >
-                  {navItem.name}
-                </button>
-              ))}
-            </nav>
           </div>
         </div>
-      </div>
 
-      {/* ——— Content Area ——— */}
-      <div className="flex-grow overflow-auto">
-        {['statistics', 'settings'].includes(currentNav.slug) ? (
-          <div className="p-4 sm:p-6 lg:p-8 bg-gray-50 dark:bg-gray-950">
-            <ActiveComponent key={currentNav.slug} />
-          </div>
-        ) : (
-          // InspectrApp takes care of its own layout now
-          <ActiveComponent key="inspectr" />
-        )}
+        {/* ——— Content Area ——— */}
+        <div className="flex-grow overflow-auto">
+          {['statistics', 'settings'].includes(currentNav.slug) ? (
+            <div className="p-4 sm:p-6 lg:p-8 bg-gray-50 dark:bg-gray-950">
+              <ActiveComponent key={currentNav.slug} />
+            </div>
+          ) : (
+            // InspectrApp takes care of its own layout now
+            <InspectrApp key="inspectr" />
+          )}
+        </div>
+
+        {/* Toast Notification from context */}
+        <ToastNotificationFromContext />
       </div>
-    </div>
+    </InspectrProvider>
+  );
+};
+
+// Component to display toast notifications from context
+const ToastNotificationFromContext = () => {
+  const { toast, setToast } = useInspectr();
+
+  if (!toast) return null;
+
+  return (
+    <ToastNotification
+      message={toast.message}
+      subMessage={toast.subMessage}
+      type={toast.type}
+      onClose={() => setToast(null)}
+    />
   );
 }
