+++ src/context/InspectrContext.jsx
@@ -1,0 +1,310 @@
+import React, { createContext, useState, useContext, useRef, useEffect } from 'react';
+import { registerApi, getConfigApi } from '../utils/inspectrSdk';
+
+// Create the context with default values
+const InspectrContext = createContext({
+  // Settings
+  apiEndpoint: '/api',
+  setApiEndpoint: () => {},
+
+  // Connection status
+  connectionStatus: 'disconnected', // 'connected' | 'reconnecting' | 'disconnected'
+  setConnectionStatus: () => {},
+
+  // Registration details
+  sseEndpoint: '',
+  channelCode: '',
+  setChannelCode: () => {},
+  channel: '',
+  setChannel: () => {},
+  token: '',
+
+  // Refs
+  reRegistrationFailedRef: { current: false },
+  userInitiatedRegistrationRef: { current: false },
+
+  // Registration functions
+  handleRegister: () => {},
+  attemptReRegistration: () => {},
+
+  // Toast notifications
+  toast: null,
+  setToast: () => {},
+});
+
+// Custom hook to use the context
+export const useInspectr = () => useContext(InspectrContext);
+
+// Provider component
+export const InspectrProvider = ({ children }) => {
+  // Settings state
+  const [apiEndpoint, setApiEndpoint] = useState(() => {
+    // Initialize from localStorage if available, otherwise use default
+    if (typeof window !== 'undefined') {
+      const storedApiEndpoint = localStorage.getItem('apiEndpoint');
+      // Normalize the endpoint by removing trailing slashes
+      return storedApiEndpoint ? storedApiEndpoint.replace(/\/+$/, '') : '/api';
+    }
+    return '/api';
+  });
+  const [connectionStatus, setConnectionStatus] = useState('disconnected');
+  const [sseEndpoint, setSseEndpoint] = useState('');
+  const [channelCode, setChannelCode] = useState('');
+  const [channel, setChannel] = useState('');
+  const [token, setToken] = useState('');
+  const [expires, setExpires] = useState('');
+  const [isInitialized, setIsInitialized] = useState(false);
+  const [toast, setToast] = useState(null);
+
+  // SSE connection reference
+  const registrationRetryCountRef = useRef(0);
+  const reRegistrationFailedRef = useRef(false);
+  const maxRegistrationRetries = 6;
+  const retryDelay = 5000; // milliseconds
+
+  const isLocalhost = typeof window !== 'undefined' && window.location.hostname === 'localhost';
+  const debugMode = typeof window !== 'undefined' && localStorage.getItem('debug') === 'true';
+
+  if (debugMode) {
+    console.log('[Inspectr] Debug Mode enabled');
+  }
+
+  // Load credentials from URL query parameters
+  const loadCredentialsFromQueryParams = () => {
+    const urlParams = new URLSearchParams(window.location.search);
+    const queryChannelCode = urlParams.get('channelCode');
+    const queryChannel = urlParams.get('channel');
+    const queryToken = urlParams.get('token');
+
+    if (queryChannelCode || queryChannel || queryToken) {
+      console.log('🔍 Found credentials in query params');
+      if (queryChannelCode) {
+        setChannelCode(queryChannelCode);
+        localStorage.setItem('channelCode', queryChannelCode);
+      }
+      if (queryChannel) {
+        setChannel(queryChannel);
+        localStorage.setItem('channel', queryChannel);
+      }
+      if (queryToken) {
+        setToken(queryToken);
+        localStorage.setItem('token', queryToken);
+      }
+      // Update the URL without reloading the page.
+      window.history.replaceState(
+        {},
+        '',
+        `${window.location.pathname}${urlParams.toString() ? '?' + urlParams.toString() : ''}`
+      );
+      return true;
+    }
+    return false;
+  };
+
+  // Load credentials from local storage
+  const loadCredentialsFromLocalStorage = () => {
+    const storedChannelCode = localStorage.getItem('channelCode');
+    const storedChannel = localStorage.getItem('channel');
+    const storedToken = localStorage.getItem('token');
+
+    if (storedChannelCode && storedChannel && storedToken) {
+      console.log('✅ Using stored credentials from localStorage');
+      setChannelCode(storedChannelCode);
+      setChannel(storedChannel);
+      setToken(storedToken);
+      return true;
+    }
+    return false;
+  };
+
+  // Load credentials from REST API
+  const loadCredentialsFromApi = async () => {
+    console.log('🔄 Fetching /app/config (Localhost, No credentials found)');
+    try {
+      const result = await getConfigApi();
+      if (result?.token && result?.sse_endpoint && result?.channel_code) {
+        console.log('✅ Loaded from /app/config:', result);
+        setChannelCode(result.channel_code);
+        localStorage.setItem('channelCode', result.channel_code);
+        setChannel(result.channel);
+        localStorage.setItem('channel', result.channel);
+        setToken(result.token);
+        localStorage.setItem('token', result.token);
+        setSseEndpoint(result.sse_endpoint);
+        localStorage.setItem('sseEndpoint', result.sse_endpoint);
+        localStorage.setItem('ingressEndpoint', result.ingress_endpoint);
+        localStorage.setItem('proxyEndpoint', result.proxy_endpoint);
+        localStorage.setItem('expose', result.expose ? 'true' : 'false');
+      }
+    } catch (err) {
+      console.error('❌ Failed to load /app/config:', err);
+    }
+  };
+
+  // Registration handler.
+  const handleRegister = async (
+    newChannelCode = channelCode,
+    newChannel = channel,
+    newToken = token,
+    showNotification
+  ) => {
+    try {
+      // Construct request body
+      let requestBody = {};
+      if (newChannelCode && newChannel) {
+        requestBody = { channel: newChannel, channel_code: newChannelCode };
+      } else if (newToken) {
+        requestBody = { token: newToken };
+      }
+
+      console.log('📤 Registering with:', requestBody);
+
+      // Register App with Inspectr
+      const result = await registerApi(apiEndpoint, requestBody);
+
+      if (result?.token && result?.sse_endpoint && result?.channel_code) {
+        console.log('✅ Registration successful');
+        registrationRetryCountRef.current = 0;
+        reRegistrationFailedRef.current = false;
+
+        setChannelCode(result.channel_code);
+        localStorage.setItem('channelCode', result.channel_code);
+        setChannel(result.channel);
+        localStorage.setItem('channel', result.channel);
+        setToken(result.token);
+        localStorage.setItem('token', result.token);
+        setExpires(result.expires);
+        localStorage.setItem('expires', result.expires);
+        setSseEndpoint(result.sse_endpoint);
+        localStorage.setItem('sseEndpoint', result.sse_endpoint);
+        localStorage.setItem('ingressEndpoint', result.ingress_endpoint);
+        localStorage.setItem('proxyEndpoint', result.proxy_endpoint);
+        localStorage.setItem('expose', result.expose ? 'true' : 'false');
+
+        if (showNotification) {
+          setToast({
+            message: 'Registration Successful',
+            subMessage: 'Your channel and access code have been registered.'
+          });
+        }
+        // Update the connection status.
+        setConnectionStatus('connected');
+        return true;
+      } else {
+        console.error('❌ Registration failed:', result);
+        setToast({
+          message: 'Registration Failed',
+          subMessage: 'Please check your channel and access code.',
+          type: 'error'
+        });
+        throw new Error('Registration failed');
+      }
+    } catch (error) {
+      console.error('❌ Registration error:', error);
+      setToast({
+        message: 'Registration Error',
+        subMessage: 'An error occurred during registration.',
+        type: 'error'
+      });
+      throw error;
+    }
+  };
+
+  const attemptReRegistration = () => {
+    setConnectionStatus('reconnecting');
+    if (registrationRetryCountRef.current < maxRegistrationRetries) {
+      registrationRetryCountRef.current += 1;
+      console.log(
+        `🔄 Attempting re-registration (${registrationRetryCountRef.current}/${maxRegistrationRetries}) in ${
+          retryDelay / 1000
+        } seconds...`
+      );
+      setTimeout(async () => {
+        try {
+          await handleRegister();
+        } catch (error) {
+          console.log('❌ Re-registration attempt failed:', error);
+          // Try again recursively.
+          attemptReRegistration();
+        }
+      }, retryDelay);
+    } else {
+      console.log(`❌ Re-registration failed after ${maxRegistrationRetries} attempts. Giving up.`);
+      registrationRetryCountRef.current = 0;
+      reRegistrationFailedRef.current = true; // Prevent further attempts
+      // Set final connection status to "disconnected"
+      setConnectionStatus('disconnected');
+    }
+  };
+
+  /**
+   * 🏁 Step 1: Load credentials from localStorage on mount
+   */
+  useEffect(() => {
+    const loadCredentials = async () => {
+      if (typeof window === 'undefined') return;
+
+      if (!loadCredentialsFromQueryParams()) {
+        if (!loadCredentialsFromLocalStorage() && isLocalhost) {
+          await loadCredentialsFromApi();
+        }
+      }
+
+      setIsInitialized(true);
+    };
+
+    loadCredentials();
+  }, []);
+
+  /**
+   * 🏁 Step 2: When `isInitialized` is true, register using stored credentials
+   * We use a ref to track if registration was triggered by the user to avoid loops
+   */
+  const userInitiatedRegistrationRef = useRef(false);
+
+  useEffect(() => {
+    if (!isInitialized) return;
+
+    // Skip auto-registration if it was triggered by the user via SettingsPanel
+    if (userInitiatedRegistrationRef.current) {
+      userInitiatedRegistrationRef.current = false;
+      return;
+    }
+
+    console.log('🔄 Auto-registering with stored credentials:', channel, channelCode);
+
+    if (channel && channelCode) {
+      handleRegister(channelCode, channel);
+    } else {
+      console.log('⚠️ Missing credentials for auto-registration, skipping.');
+    }
+  }, [isInitialized, channel, channelCode]);
+
+  // Context value
+  const contextValue = {
+    apiEndpoint,
+    setApiEndpoint,
+    connectionStatus,
+    setConnectionStatus,
+    sseEndpoint,
+    channelCode,
+    setChannelCode,
+    channel,
+    setChannel,
+    token,
+    reRegistrationFailedRef,
+    userInitiatedRegistrationRef,
+    handleRegister,
+    attemptReRegistration,
+    toast,
+    setToast,
+  };
+
+  return (
+    <InspectrContext.Provider value={contextValue}>
+      {children}
+    </InspectrContext.Provider>
+  );
+};
+
+export default InspectrContext;
